chore: Rename clean to db clean
