+++ chroma_ops/db.py
@@ -1,6 +1,7 @@
 import typer
 
 from chroma_ops.db_info import command as db_info_command
+from chroma_ops.db_clean import command as db_clean_command
 
 db_commands = typer.Typer(no_args_is_help=True)
 
@@ -10,3 +11,9 @@ db_commands.command(
     "Useful to understand how your Chroma works or get support from the team.",
     no_args_is_help=True,
 )(db_info_command)
+
+db_commands.command(
+    name="clean",
+    help="Clean up orphanated HNSW segment subdirectories.",
+    no_args_is_help=True,
+)(db_clean_command)
