+++ chroma_ops/main.py
@@ -3,7 +3,6 @@ import typer
 from chroma_ops.fts import fts_commands
 from chroma_ops.wal import wal_commands
 from chroma_ops.db import db_commands
-from chroma_ops.clean import command as clean_command
 from chroma_ops.hnsw import hnsw_commands
 from chroma_ops.collection import collection_commands
 
@@ -12,12 +11,6 @@ app = typer.Typer(no_args_is_help=True, help="ChromaDB Ops Commands.")
 
 app.add_typer(wal_commands, name="wal", help="WAL maintenance commands")
 
-
-app.command(
-    name="clean",
-    help="Clean up orphaned vector segment directories.",
-    no_args_is_help=True,
-)(clean_command)
 
 app.add_typer(db_commands, name="db", help="DB maintenance commands")
